Updated language processing and page rendering
