+++ public/languages/en.json
@@ -1,0 +1,6 @@
+{
+	"/": {
+		"hello_world": "Hello world!",
+		"goodbye_world": "Goodbye world!"
+	}
+}
