+++ public/languages/ja.json
@@ -1,0 +1,6 @@
+{
+	"/": {
+		"hello_world": "おはよう！",
+		"goodbye_world": "またね！"
+	}
+}
