--- public/views/mobile/en/footer.ejs
@@ -1,5 +1,0 @@
-<footer>
-  <div class="footer-inner">
-  </div>
-</footer>
-<script src="/js/bundle.min.js" nonce="<%= nonce %>"></script>
