--- public/views/mobile/en/head.ejs
@@ -1,10 +1,0 @@
-<meta charset="UTF-8" />
-<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-<link rel="preconnect" href="https://fonts.googleapis.com" />
-<link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-<link
-  href="https://fonts.googleapis.com/css2?family=Barlow+Condensed&display=swap"
-  rel="stylesheet"
-/>
-<title>NodeJs Template | <%= title %></title>
-<link type="image/x-icon" rel="shortcut icon" href="/img/favicon.ico" />
