--- public/views/mobile/en/header.ejs
@@ -1,2 +1,0 @@
-<header>
-</header>
