--- public/views/mobile/en/home.ejs
@@ -1,11 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-	<head>
-		<%- include('./head'); %>
-	</head>
-	<body>
-		<%- include('./header'); %>
-		Hello world!
-		<%- include('./footer'); %>
-	</body>
-</html>
