--- public/views/mobile/ja/header.ejs
@@ -1,2 +1,0 @@
-<header>
-</header>
